% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_names}
\alias{clean_names}
\title{Remove municipal status indicators from non-duplicated region names. Indicators are retained for
regions that would otherwise have duplicated names.}
\usage{
clean_names(dfr)
}
\arguments{
\item{dfr}{a cancensus data frame with region names}
}
\description{
Remove municipal status indicators from non-duplicated region names. Indicators are retained for
regions that would otherwise have duplicated names.
}
