% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{common_cts}
\alias{common_cts}
\title{Aggregate variables to common cts, returns data2 on new tiling matching data1 geography}
\usage{
common_cts(data1, data2, data2_sum_vars)
}
\arguments{
\item{data1}{Cancensus CT level datatset for year1 < year2 to serve as base for common geography}

\item{data2}{Cancensus CT level datatset for year2 to be aggregated to common geography}

\item{data2_sum_vars}{vector of variable names to by summed up when aggregating geographies}
}
\description{
Aggregate variables to common cts, returns data2 on new tiling matching data1 geography
}
