% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_helpers.R
\name{ensure_local_statcan_xtab}
\alias{ensure_local_statcan_xtab}
\title{Download csv xtab data from url}
\usage{
ensure_local_statcan_xtab(code, url, refresh = FALSE,
  system_unzip = TRUE)
}
\description{
Download csv xtab data from url
}
