% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_cov_variables}
\alias{find_cov_variables}
\title{Convenience function to serach for census variable}
\usage{
find_cov_variables(data, search_string)
}
\description{
Convenience function to serach for census variable
}
