% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{geom_waffle}
\alias{geom_waffle}
\title{waffle geometry for ggplot}
\usage{
geom_waffle(color = "white", size = 0.4, ...)
}
\arguments{
\item{color}{background color}

\item{size}{size of grid}
}
\description{
waffle geometry for ggplot
}
