% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_age_data}
\alias{get_age_data}
\title{Returns dataset with age data for the specified region}
\usage{
get_age_data(dataset, regions, should_sum = FALSE, refresh = FALSE)
}
\arguments{
\item{dataset}{Cancensus Datatset}

\item{regions}{Cancensus Regions}

\item{should_sum}{logical, should the regions be summed up or reported separately}

\item{refresh}{logical, should cached data be refreshed}
}
\description{
Returns dataset with age data for the specified region
}
