% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_cov_census_data}
\alias{get_cov_census_data}
\title{load and parse census data for a given year}
\usage{
get_cov_census_data(year, use_cache = TRUE)
}
\description{
load and parse census data for a given year
}
