% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_helpers.R
\name{get_multi_header_xtab}
\alias{get_multi_header_xtab}
\title{read in beyond20/20 export of statcan xtab into tidy long-format data frame}
\usage{
get_multi_header_xtab(path, parse_geography_column = function(data) {
    data
})
}
\description{
read in beyond20/20 export of statcan xtab into tidy long-format data frame
}
