% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_old_toronto_data}
\alias{get_old_toronto_data}
\title{Get variables for CTs making up Old Toronto}
\usage{
get_old_toronto_data(dataset, vectors = c(), labels = "short",
  geo_format = NA, also_new_toronto = FALSE, aggregate = FALSE)
}
\description{
Get variables for CTs making up Old Toronto
}
