% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_shapefile}
\alias{get_shapefile}
\title{download zipped shapefile and read shapefile.}
\usage{
get_shapefile(path, file_mask = NA)
}
\arguments{
\item{path}{URL string to zipped shape file}

\item{file_mask}{optional grep string in case there are several shape files in the package}
}
\value{
an sf object with the data from the shape file
}
\description{
download zipped shapefile and read shapefile.
}
