% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_helpers.R
\name{get_unique_values_for_field}
\alias{get_unique_values_for_field}
\title{Get unique values for field}
\usage{
get_unique_values_for_field(connection, table_name, field,
  where_string = NA)
}
\arguments{
\item{connection}{Database connection}

\item{table_name}{Name of the table}

\item{field}{name of the field}
}
\description{
Get unique values for field
}
