% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_vector_tiles}
\alias{get_vector_tiles}
\title{Get vector tile data, expects option variable nextzen_API_key to be set}
\usage{
get_vector_tiles(bbox)
}
\arguments{
\item{bbox}{bounding box for which to get vector tile data}
}
\description{
Get vector tile data, expects option variable nextzen_API_key to be set
}
