% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{map_theme}
\alias{map_theme}
\title{helper for clean map theme for sf}
\format{An object of class \code{list} of length 2.}
\usage{
map_theme
}
\description{
helper for clean map theme for sf
}
\keyword{datasets}
