% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_helpers.R
\name{parse_geography_from}
\alias{parse_geography_from}
\title{parse geography name column from beyond20/20 export˚}
\usage{
parse_geography_from(data, geo_column = "Geography")
}
\description{
parse geography name column from beyond20/20 export˚
}
