% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_named_vector}
\alias{to_named_vector}
\title{turn two columns of a tibble into a named vector}
\usage{
to_named_vector(data, name_column, value_column)
}
\arguments{
\item{data}{a tibble}

\item{name_column}{string with the name of the column to be used as names in the vector}

\item{value_column}{string with the name of the column to be used as values in the vector}
}
\value{
a named vector
}
\description{
turn two columns of a tibble into a named vector
}
