% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{waffle_tile}
\alias{waffle_tile}
\title{generate data for waffle plots`}
\usage{
waffle_tile(
  data,
  grouping_variables,
  category = "Type",
  value = "Value",
  nrow = 10,
  ncol = 10
)
}
\arguments{
\item{data}{tibble with data}

\item{grouping_variables}{variables used for grouping}

\item{category}{column name for coloring}

\item{nrow}{number of columns in waffle}

\item{vakue}{column name that contains the counts}
}
\description{
generate data for waffle plots`
}
