% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xml_xtab_for}
\alias{xml_xtab_for}
\title{Download xml xtab data from url, tag with code for caching
useful for older (2006) data that does not come as csv option}
\usage{
xml_xtab_for(code, url, refresh = FALSE)
}
\description{
Download xml xtab data from url, tag with code for caching
useful for older (2006) data that does not come as csv option
}
