% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_helpers.R
\name{xtab_for}
\alias{xtab_for}
\title{Download csv xtab data from url, tag with code for caching}
\usage{
xtab_for(code, url = NA, refresh = FALSE, system_unzip = FALSE)
}
\description{
Download csv xtab data from url, tag with code for caching
}
