% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_age_data}
\alias{get_age_data}
\title{Returns dataset with age data for the specified region}
\usage{
get_age_data(dataset, regions, sum = FALSE)
}
\arguments{
\item{dataset}{Cancensus Datatset}

\item{regions}{Cancensus Regions}
}
\description{
Returns dataset with age data for the specified region
}
